+++ packages/pretty-format/src/__tests__/prettyFormat.test.ts
@@ -252,28 +252,25 @@ describe('prettyFormat()', () => {
     expect(prettyFormat(val)).toEqual('-0');
   });
 
-  /* global BigInt */
-  if (typeof BigInt === 'function') {
-    it('prints a positive bigint', () => {
-      const val = BigInt(123);
-      expect(prettyFormat(val)).toEqual('123n');
-    });
+  it('prints a positive bigint', () => {
+    const val = BigInt(123);
+    expect(prettyFormat(val)).toEqual('123n');
+  });
 
-    it('prints a negative bigint', () => {
-      const val = BigInt(-123);
-      expect(prettyFormat(val)).toEqual('-123n');
-    });
+  it('prints a negative bigint', () => {
+    const val = BigInt(-123);
+    expect(prettyFormat(val)).toEqual('-123n');
+  });
 
-    it('prints zero bigint', () => {
-      const val = BigInt(0);
-      expect(prettyFormat(val)).toEqual('0n');
-    });
+  it('prints zero bigint', () => {
+    const val = BigInt(0);
+    expect(prettyFormat(val)).toEqual('0n');
+  });
 
-    it('prints negative zero bigint', () => {
-      const val = BigInt(-0);
-      expect(prettyFormat(val)).toEqual('0n');
-    });
-  }
+  it('prints negative zero bigint', () => {
+    const val = BigInt(-0);
+    expect(prettyFormat(val)).toEqual('0n');
+  });
 
   it('prints a date', () => {
     const val = new Date(10e11);
@@ -332,10 +329,26 @@ describe('prettyFormat()', () => {
   });
 
   it('prints an object with sorted properties', () => {
-    /* eslint-disable sort-keys */
+    // eslint-disable-next-line sort-keys
     const val = {b: 1, a: 2};
-    /* eslint-enable sort-keys */
     expect(prettyFormat(val)).toEqual('Object {\n  "a": 2,\n  "b": 1,\n}');
+  });
+
+  it('prints an object with keys in their original order', () => {
+    // eslint-disable-next-line sort-keys
+    const val = {b: 1, a: 2};
+    const compareKeys = () => 0;
+    expect(prettyFormat(val, {compareKeys})).toEqual(
+      'Object {\n  "b": 1,\n  "a": 2,\n}',
+    );
+  });
+
+  it('prints an object with keys sorted in reverse order', () => {
+    const val = {a: 1, b: 2};
+    const compareKeys = (a: string, b: string) => (a > b ? -1 : 1);
+    expect(prettyFormat(val, {compareKeys})).toEqual(
+      'Object {\n  "b": 2,\n  "a": 1,\n}',
+    );
   });
 
   it('prints regular expressions from constructors', () => {