+++ scripts/cleanE2e.js
@@ -18,6 +18,7 @@ const excludedModules = [
   'e2e/presets/js-type-module/node_modules/',
   'e2e/presets/json/node_modules/',
   'e2e/presets/mjs/node_modules/',
+  'e2e/resolve-conditions/node_modules/',
 ].map(dir => normalize(dir));
 
 const e2eNodeModules = glob('e2e/*/node_modules/')